feat: add search query with filters, sorting, pagination and facets, change PagedResponse totalCount type to long

+++ src/Pereprodai.Shared/Application/DTOs/PagedResponse.cs
@@ -2,6 +2,6 @@ namespace Pereprodai.Shared.Application.DTOs;
 
 public record PagedResponse<T>(
     List<T> Items,
-    int TotalCount,
+    long TotalCount,
     int Page,
     int PageSize);
